+++ src-tauri/crates/tts-helper-audio/src/sources.rs
@@ -1,11 +1,11 @@
 mod on_finish;
 mod on_start;
-mod skip_if;
+mod skip_when;
 mod source_ext;
 mod take_for;
 
 pub use on_finish::*;
 pub use on_start::*;
-pub use skip_if::*;
+pub use skip_when::*;
 pub use source_ext::*;
 pub use take_for::*;
